Entregables_Pokemon/Roberto Chavez Garcia/baz_behave/requirements.txt
behave == 1.2.6
Appium-Python-Client==2.11.1
allure-behave==2.13.2
python-dotenv==1.0.0
selenium==4.8.3
